third_party/WebKit/LayoutTests/http/tests/payments/payment-app-manager.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../../serviceworker/resources/test-helpers.js"></script>
<script>

promise_test(test => {
    var registration;
    var script_url = 'resources/empty-worker.js';
    var scope = 'resources/';

    return service_worker_unregister_and_register(test, script_url, scope)
      .then(r => {
          registration = r;
          return wait_for_state(test, registration.installing, 'activated');
        })
      .then(state => {
          assert_equals(state, 'activated');
          return registration.paymentAppManager.setManifest({
              label: 'Payment App'
            });
        })
      .then(result => {
          unreached_fulfillment(test);
        })
      .catch(error => {
          assert_equals(error.name, 'NotSupportedError');
        });
  }, 'setManifest() should reject NotSupportedError for now');

</script>
